exercise sol

+++ src/main/java/com/example/demo/student/StudentService.java
@@ -1,7 +1,11 @@
 package com.example.demo.student;
 
+import com.example.demo.student.exception.BadRequestException;
+import com.example.demo.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.ResponseStatus;
 
 import java.util.List;
 
@@ -16,12 +20,21 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
-        // check if email is taken
+        Boolean existsEmail = studentRepository
+                .selectExistsEmail(student.getEmail());
+        if (existsEmail) {
+            throw new BadRequestException(
+                    "Email " + student.getEmail() + " taken");
+        }
+
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
-        // check if student exists
+        if(!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id " + studentId + " does not exists");
+        }
         studentRepository.deleteById(studentId);
     }
 }
